+++ apps/api/src/ticle/dto/createTicleDto.ts
@@ -1,0 +1,10 @@
+export type CreateTicleDto = {
+  speakerName: string;
+  speakerEmail: string;
+  speakerIntroduce: string;
+  title: string;
+  content: string;
+  startTime: Date;
+  endTime: Date;
+  tags?: string[];
+};
