+++ apps/api/src/ticle/ticle.controller.ts
@@ -1,5 +1,6 @@
 import { Body, Controller, Get, Param, Post, Query } from '@nestjs/common';
 
+import { CreateTicleDto } from './dto/createTicleDto';
 import { TicleService } from './ticle.service';
 
 @Controller('ticle')
@@ -24,14 +25,3 @@ export class TicleController {
   @Post(':ticleId/apply')
   applyToTicle(@Param('ticleId') params: any) {}
 }
-
-export type CreateTicleDto = {
-  speakerName: string;
-  speakerEmail: string;
-  speakerIntroduce: string;
-  title: string;
-  content: string;
-  startTime: Date;
-  endTime: Date;
-  tags?: string[];
-};
